refactor(challenge): controller 분리
