+++ src/test/java/com/sopt/cherrish/domain/challenge/recommendation/presentation/ChallengeRecommendationControllerTest.java
@@ -1,17 +1,13 @@
-package com.sopt.cherrish.domain.challenge.presentation;
+package com.sopt.cherrish.domain.challenge.recommendation.presentation;
 
 import static com.sopt.cherrish.domain.challenge.fixture.ChallengeTestFixture.emptyRoutinesRecommendation;
-import static com.sopt.cherrish.domain.challenge.fixture.ChallengeTestFixture.homecareRoutineList;
 import static com.sopt.cherrish.domain.challenge.fixture.ChallengeTestFixture.recommendationRequest;
 import static com.sopt.cherrish.domain.challenge.fixture.ChallengeTestFixture.skinMoisturizingRecommendation;
 import static com.sopt.cherrish.domain.challenge.fixture.ChallengeTestFixture.wrinkleCareRecommendation;
 import static org.mockito.BDDMockito.given;
-import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get;
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.post;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.jsonPath;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
-
-import java.util.List;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
@@ -23,14 +19,11 @@ import org.springframework.test.context.bean.override.mockito.MockitoBean;
 import org.springframework.test.web.servlet.MockMvc;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
-import com.sopt.cherrish.domain.challenge.core.application.facade.ChallengeCreationFacade;
-import com.sopt.cherrish.domain.challenge.core.presentation.ChallengeController;
-import com.sopt.cherrish.domain.challenge.homecare.application.service.HomecareRoutineService;
 import com.sopt.cherrish.domain.challenge.recommendation.application.service.AiChallengeRecommendationService;
 
-@WebMvcTest(ChallengeController.class)
-@DisplayName("ChallengeController 통합 테스트")
-class ChallengeControllerTest {
+@WebMvcTest(ChallengeRecommendationController.class)
+@DisplayName("ChallengeRecommendationController 통합 테스트")
+class ChallengeRecommendationControllerTest {
 
 	@Autowired
 	private MockMvc mockMvc;
@@ -39,50 +32,7 @@ class ChallengeControllerTest {
 	private ObjectMapper objectMapper;
 
 	@MockitoBean
-	private HomecareRoutineService homecareRoutineService;
-
-	@MockitoBean
 	private AiChallengeRecommendationService aiRecommendationService;
-
-	@MockitoBean
-	private ChallengeCreationFacade challengeCreationFacade;
-
-	@Nested
-	@DisplayName("GET /api/challenges/homecare-routines - 홈케어 루틴 목록 조회")
-	class GetHomecareRoutines {
-
-		@Test
-		@DisplayName("성공 - 루틴 목록 반환")
-		void success() throws Exception {
-			// given
-			given(homecareRoutineService.getAllHomecareRoutines()).willReturn(homecareRoutineList());
-
-			// when & then
-			mockMvc.perform(get("/api/challenges/homecare-routines"))
-				.andExpect(status().isOk())
-				.andExpect(jsonPath("$.data").isArray())
-				.andExpect(jsonPath("$.data.length()").value(6))
-				.andExpect(jsonPath("$.data[0].id").value(1))
-				.andExpect(jsonPath("$.data[0].name").value("SKIN_MOISTURIZING"))
-				.andExpect(jsonPath("$.data[0].description").value("피부 보습 관리"))
-				.andExpect(jsonPath("$.data[1].id").value(2))
-				.andExpect(jsonPath("$.data[2].id").value(3))
-				.andExpect(jsonPath("$.data[5].id").value(6));
-		}
-
-		@Test
-		@DisplayName("성공 - 빈 목록 반환")
-		void emptyList() throws Exception {
-			// given
-			given(homecareRoutineService.getAllHomecareRoutines()).willReturn(List.of());
-
-			// when & then
-			mockMvc.perform(get("/api/challenges/homecare-routines"))
-				.andExpect(status().isOk())
-				.andExpect(jsonPath("$.data").isArray())
-				.andExpect(jsonPath("$.data.length()").value(0));
-		}
-	}
 
 	@Nested
 	@DisplayName("POST /api/challenges/ai-recommendations - AI 챌린지 추천 생성")
